Rename test package from "products" to "items"
